feat: move current user store to store directory
